Fix to avoid contrail-cassandra-status zombie proc
contrail-nodemgr periodically calls contrail-cassandra-status utility,
via non-blocking subprocess.Popen() and does not take care about
its return code. As a result, when child process ends and gracefully
exits, it is still present in "ps" output as zombie process
"[contrail-cassan]".

It is not critical, as just one zombie process is present, and it is
actually exited. But since somebody may be scared and confused by this
"[contrail-cassan]", it is better just to not let this process become
zombie.

Proposed fix: use "subprocess.call()" instead of "Popen()" for
"contrail-cassandra-status". Use it with "shell=True" - to avoid
zombies, and use "&" at the end - to run in non-blocking way.
And also do this thing for "contrail-cassandra-repair" as it has same problem.

Change-Id: Ib921904c2238262ac85f45b0f19aadd2640c7de5